Added Laptopman's Creed
I’m not sure how many people will get this, but I think it’s awesome.
And it was a chance for me to have a little fun.  I’ve always thought
the Rifleman’s Creed was a respectful and still amusing way of looking
at how we use the tools that our lives depend on, and for programmers,
our laptop really does mean about that much. At least to me.